155150201111334_I GEDE ADI BRAHMAN NUGRAHA
REVISI
